+++ apis/v1alpha1/doc.go
@@ -4,6 +4,6 @@ Copyright 2021 Upbound Inc.
 
 // Package v1alpha1 contains the core resources of the openstack jet provider.
 // +kubebuilder:object:generate=true
-// +groupName=openstack.upbound.io
+// +groupName=openstack.crossplane.io
 // +versionName=v1alpha1
 package v1alpha1